refactor(core): separate resource api file to two files due to file size
