+++ packages/core/src/routes/resource.ts
@@ -1,6 +1,6 @@
 import { isManagementApi, Resources, Scopes } from '@logto/schemas';
 import { generateStandardId } from '@logto/shared';
-import { tryThat, yes } from '@silverhand/essentials';
+import { yes } from '@silverhand/essentials';
 import { boolean, object, string } from 'zod';
 
 import RequestError from '#src/errors/RequestError/index.js';
@@ -9,7 +9,6 @@ import koaPagination from '#src/middleware/koa-pagination.js';
 import koaQuotaGuard from '#src/middleware/koa-quota-guard.js';
 import assertThat from '#src/utils/assert-that.js';
 import { attachScopesToResources } from '#src/utils/resource.js';
-import { parseSearchParamsForSearch } from '#src/utils/search.js';
 
 import type { AuthedRouter, RouterInitArgs } from './types.js';
 
@@ -34,14 +33,6 @@ export default function resourceRoutes<T extends AuthedRouter>(
       deleteResourceById,
     },
     scopes: scopeQueries,
-    scopes: {
-      countScopesByResourceId,
-      deleteScopeById,
-      searchScopesByResourceId,
-      findScopeByNameAndResourceId,
-      insertScope,
-      updateScopeById,
-    },
   } = queries;
 
   router.get(
@@ -194,160 +185,6 @@ export default function resourceRoutes<T extends AuthedRouter>(
       );
 
       await deleteResourceById(id);
-      ctx.status = 204;
-
-      return next();
-    }
-  );
-
-  router.get(
-    '/resources/:resourceId/scopes',
-    koaPagination(),
-    koaGuard({
-      params: object({ resourceId: string().min(1) }),
-      status: [200, 400],
-      response: Scopes.guard.array(),
-    }),
-    async (ctx, next) => {
-      const {
-        params: { resourceId },
-      } = ctx.guard;
-      const { limit, offset } = ctx.pagination;
-      const { searchParams } = ctx.request.URL;
-
-      return tryThat(
-        async () => {
-          const search = parseSearchParamsForSearch(searchParams);
-
-          const [{ count }, scopes] = await Promise.all([
-            countScopesByResourceId(resourceId, search),
-            searchScopesByResourceId(resourceId, search, limit, offset),
-          ]);
-
-          // Return totalCount to pagination middleware
-          ctx.pagination.totalCount = count;
-          ctx.body = scopes;
-
-          return next();
-        },
-        (error) => {
-          if (error instanceof TypeError) {
-            throw new RequestError(
-              { code: 'request.invalid_input', details: error.message },
-              error
-            );
-          }
-          throw error;
-        }
-      );
-    }
-  );
-
-  router.post(
-    '/resources/:resourceId/scopes',
-    koaGuard({
-      params: object({ resourceId: string().min(1) }),
-      body: Scopes.createGuard.pick({ name: true, description: true }),
-      response: Scopes.guard,
-      status: [201, 422, 400, 404],
-    }),
-    async (ctx, next) => {
-      const {
-        params: { resourceId },
-        body,
-      } = ctx.guard;
-
-      await quota.guardKey('scopesPerResourceLimit', resourceId);
-
-      assertThat(!/\s/.test(body.name), 'scope.name_with_space');
-
-      assertThat(
-        await findResourceById(resourceId),
-        new RequestError({
-          code: 'entity.not_exists_with_id',
-          name: 'resource',
-          id: resourceId,
-          resourceId,
-          status: 404,
-        })
-      );
-
-      assertThat(
-        !(await findScopeByNameAndResourceId(body.name, resourceId)),
-        new RequestError({
-          code: 'scope.name_exists',
-          name: body.name,
-          status: 422,
-        })
-      );
-
-      ctx.status = 201;
-      ctx.body = await insertScope({
-        ...body,
-        id: generateStandardId(),
-        resourceId,
-      });
-
-      return next();
-    }
-  );
-
-  router.patch(
-    '/resources/:resourceId/scopes/:scopeId',
-    koaGuard({
-      params: object({ resourceId: string().min(1), scopeId: string().min(1) }),
-      body: Scopes.createGuard.pick({ name: true, description: true }).partial(),
-      response: Scopes.guard,
-      status: [200, 400, 404, 422],
-    }),
-    async (ctx, next) => {
-      const {
-        params: { scopeId, resourceId },
-        body,
-      } = ctx.guard;
-
-      assertThat(
-        await findResourceById(resourceId),
-        new RequestError({
-          code: 'entity.not_exists_with_id',
-          name: 'resource',
-          id: resourceId,
-          resourceId,
-          status: 404,
-        })
-      );
-
-      if (body.name) {
-        assertThat(!/\s/.test(body.name), 'scope.name_with_space');
-        assertThat(
-          !(await findScopeByNameAndResourceId(body.name, resourceId, scopeId)),
-          new RequestError({
-            code: 'scope.name_exists',
-            name: body.name,
-            status: 422,
-          })
-        );
-      }
-
-      ctx.body = await updateScopeById(scopeId, body);
-
-      return next();
-    }
-  );
-
-  router.delete(
-    '/resources/:resourceId/scopes/:scopeId',
-    koaGuard({
-      params: object({ resourceId: string().min(1), scopeId: string().min(1) }),
-      status: [204, 404],
-    }),
-    async (ctx, next) => {
-      const {
-        params: { scopeId },
-      } = ctx.guard;
-
-      await deleteScopeById(scopeId);
-
       ctx.status = 204;
 
       return next();
